% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{name}
\alias{name}
\title{Meu Primeiro Pacote R}
\usage{
name(x)
}
\arguments{
\item{x}{Seu nome}
}
\value{
Seu nome em uma frase.
}
\description{
Uma funcao que cria uma mensagem personalizada
}
\details{
Sempre importante lembrar que seu nome e do tipo character,
portanto, deve ficar entre aspas.
}
\examples{
name("Felipe")
}
\author{
Felipe Rocha
}
